Metadata engine: overhaul internals to preserve floating-point resolutions
Relates to #260.

PD exposes image resolution as floating-point in various dialogs (Image > Resize, Canvas Resize, Content-Aware Resize), but internally, a number of places cast those floating-point values to integers.  PhotoShop internally supports the notion of floating-point DPIs, so we may as well too.

As of this commit, all of PD's internal DPI code is now safe for floating-point data.

As part of this overhaul, I've modified the way we read DPI from files (again).  A hierarchy is now used for retrieving metadata:

- If a Photoshop-specific DPI segment is found, we try to read it first.  We do this because PS only writes valid values to its own metadata segment; traditional segments - like EXIF or JFIF - may receive totally different values from what the user sees on-screen.
- If an EXIF segment is found, we use it next.  EXIF has been the primary interop method for 20+ years now, and all photo editing software *should* support it.
- ...But, if need be, we will also perform a final check for any format-specific resolution metadata.  In JPEGs this may be JFIF; in PNGs, a pHYs chunk.  It varies.  We're not picky about *where* it's found - as long as it looks like valid resolution data, we'll accept it.

I think this hierarchy makes the most sense for modern images.  I always have mixed feelings about "playing nicely" with badly formatted Photoshop files, but because PD should "just work" for the average user, I think this is an okay trade-off.  It's always frustrating when Photoshop appears to save and load data correctly, but other programs don't - so even though we're not at fault, we'll try to work around Adobe's bullshit... again.

+++ Modules/FileMenu.bas
@@ -412,7 +412,7 @@ Public Function CreateNewImage(Optional ByVal newImageParameters As String)
     Set cParams = New pdParamXML
     cParams.SetParamString newImageParameters
     
-    Dim newWidth As Long, newHeight As Long, newDPI As Long
+    Dim newWidth As Long, newHeight As Long, newDPI As Double
     Dim newBackgroundType As Long, newBackgroundColor As Long
     
     With cParams
